XIVY-18772 fixes for xhtml validation
