+++ engine-cockpit/src/ch/ivyteam/enginecockpit/monitor/monitor/Monitor.java
@@ -14,7 +14,6 @@ import java.util.Optional;
 import java.util.stream.Collectors;
 
 import org.primefaces.model.charts.ChartData;
-import org.primefaces.model.charts.ChartModel;
 import org.primefaces.model.charts.axes.cartesian.CartesianScaleTitle;
 import org.primefaces.model.charts.axes.cartesian.CartesianScales;
 import org.primefaces.model.charts.axes.cartesian.linear.CartesianLinearAxes;
@@ -164,7 +163,7 @@ public class Monitor {
     return series;
   }
 
-  public ChartModel getModel() {
+  public LineChartModel getModel() {
     calcNewValues();
     return model;
   }
