+++ engine-cockpit/src/ch/ivyteam/enginecockpit/security/model/MemberProperty.java
@@ -1,23 +1,20 @@
 package ch.ivyteam.enginecockpit.security.model;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 import javax.faces.application.FacesMessage;
 import javax.faces.context.FacesContext;
 
 import ch.ivyteam.enginecockpit.system.ManagerBean;
 import ch.ivyteam.ivy.environment.Ivy;
-import ch.ivyteam.ivy.security.IRole;
-import ch.ivyteam.ivy.security.IUser;
 
 public class MemberProperty {
   private String memberName;
-  private List<SecurityMemberProperty> properties;
+  protected List<SecurityMemberProperty> properties;
   private List<SecurityMemberProperty> filteredProperties;
-  private SecurityMemberProperty property;
+  protected SecurityMemberProperty property;
   private String filter;
-  private final ManagerBean managerBean;
+  final ManagerBean managerBean;
 
   public MemberProperty() {
     var context = FacesContext.getCurrentInstance();
@@ -74,70 +71,4 @@ public class MemberProperty {
     FacesContext.getCurrentInstance().addMessage("propertiesMessage",
         new FacesMessage(Ivy.cm().co("/memberProperties/RemovePropertyMessage"), ""));
   }
-
-  public class RoleProperty extends MemberProperty {
-    private IRole role;
-
-    @Override
-    public void setMemberName(String memberName) {
-      super.setMemberName(memberName);
-      role = managerBean.getSelectedSecuritySystem().getSecurityContext().roles().find(memberName);
-      reloadProperties();
-    }
-
-    private void reloadProperties() {
-      super.properties = role.getAllPropertyNames().stream()
-          .map(key -> new SecurityMemberProperty(key, role.getProperty(key), false))
-          .collect(Collectors.toList());
-    }
-
-    public void saveProperty() {
-      role.setProperty(super.property.getKey(), super.property.getValue());
-      super.savePropertyMessage();
-      reloadProperties();
-    }
-
-    public void removeProperty(String propertyName) {
-      role.removeProperty(propertyName);
-      super.removePropertyMessage();
-      reloadProperties();
-    }
-  }
-
-  public class UserProperty extends MemberProperty {
-    private IUser user;
-
-    @Override
-    public void setMemberName(String memberName) {
-      super.setMemberName(memberName);
-      user = managerBean.getSelectedSecuritySystem().getSecurityContext().users().find(memberName);
-      reloadProperties();
-    }
-
-    private void reloadProperties() {
-      super.properties = user.getAllPropertyNames().stream()
-          .map(key -> new SecurityMemberProperty(key, user.getProperty(key), user.isPropertyBacked(key)))
-          .collect(Collectors.toList());
-    }
-
-    public void saveProperty() {
-      if (user.isPropertyBacked(super.property.getKey())) {
-        FacesContext.getCurrentInstance().addMessage("propertiesMessage",
-            new FacesMessage(FacesMessage.SEVERITY_ERROR, Ivy.cm().co("/common/Error"),
-                Ivy.cm().content("/memberProperties/SavePropertyErrorMessage")
-                    .replace("property", super.property.getKey()).get()));
-        return;
-      }
-      user.setProperty(super.property.getKey(), super.property.getValue());
-      super.savePropertyMessage();
-      reloadProperties();
-    }
-
-    public void removeProperty(String propertyName) {
-      user.removeProperty(propertyName);
-      super.removePropertyMessage();
-      reloadProperties();
-    }
-  }
-
 }
