+++ engine-cockpit/src/ch/ivyteam/enginecockpit/services/DatabaseDetailBean.java
@@ -19,9 +19,7 @@ import ch.ivyteam.db.jdbc.JdbcDriver;
 import ch.ivyteam.enginecockpit.commons.Property;
 import ch.ivyteam.enginecockpit.commons.ResponseHelper;
 import ch.ivyteam.enginecockpit.monitor.mbeans.ivy.DatabaseMonitor;
-import ch.ivyteam.enginecockpit.services.help.HelpServices;
 import ch.ivyteam.enginecockpit.services.model.ConnectionTestResult;
-import ch.ivyteam.enginecockpit.services.model.ConnectionTestResult.IConnectionTestResult;
 import ch.ivyteam.enginecockpit.services.model.ConnectionTestResult.TestResult;
 import ch.ivyteam.enginecockpit.services.model.ConnectionTestWrapper;
 import ch.ivyteam.enginecockpit.services.model.DatabaseDto;
@@ -39,7 +37,7 @@ import ch.ivyteam.ivy.environment.Ivy;
 
 @ManagedBean
 @ViewScoped
-public class DatabaseDetailBean extends HelpServices implements IConnectionTestResult, PropertyEditor {
+public class DatabaseDetailBean extends DetailView {
 
   private DatabaseDto database;
   private List<ExecStatement> history;
@@ -131,6 +129,7 @@ public class DatabaseDetailBean extends HelpServices implements IConnectionTestR
     reloadExternalDb();
   }
 
+  @Override
   public void removeProperty(String name) {
     databases.remove(database.getName() + "." + "Properties" + "." + name);
     reloadExternalDb();
