+++ engine-cockpit/src/ch/ivyteam/enginecockpit/services/FeatureEditor.java
@@ -15,6 +15,8 @@ public interface FeatureEditor {
 
   void saveFeature(boolean isNewFeature);
 
+  public void removeFeature(String name);
+
   default Feature findFeature(String clazz) {
     return getFeatures().stream()
         .filter(feature -> feature.getClazz().equals(clazz))
