+++ engine-cockpit/webContent/includes/components/security/member-properties.xhtml
@@ -3,6 +3,7 @@
   xmlns:c="http://xmlns.jcp.org/jsp/jstl/core" xmlns:ic="http://ivyteam.ch/jsf/component"
   xmlns:p="http://primefaces.org/ui" xmlns:pe="http://primefaces.org/ui/extensions">
 
+<!-- memberProperty: ch.ivyteam.enginecockpit.security.model.RoleProperty -->
 <div class="card">
   <h:form id="propertiesForm">
     <p:growl for="propertiesMessage" id="propertiesMessage" showDetail="true" />
