+++ engine-cockpit/webContent/includes/components/services/notification-channels-table.xhtml
@@ -3,6 +3,7 @@
   xmlns:c="http://xmlns.jcp.org/jsp/jstl/core" xmlns:ic="http://ivyteam.ch/jsf/component"
   xmlns:p="http://primefaces.org/ui" xmlns:pe="http://primefaces.org/ui/extensions"
   xmlns:cc="http://xmlns.jcp.org/jsf/composite/composite">
+<!-- app: ch.ivyteam.enginecockpit.application.model.Application-->
 
 <h:form id="tableForm">
   <p:dataTable paginator="true" rows="20" paginatorPosition="bottom" paginatorAlwaysVisible="false"
