+++ engine-cockpit/webContent/includes/dialogs/deployment.xhtml
@@ -2,6 +2,8 @@
   xmlns:h="http://xmlns.jcp.org/jsf/html" xmlns:ui="http://xmlns.jcp.org/jsf/facelets"
   xmlns:cc="http://xmlns.jcp.org/jsf/composite" xmlns:ic="http://ivyteam.ch/jsf/component"
   xmlns:p="http://primefaces.org/ui">
+  <!-- appName: java.lang.String -->
+  <!-- appVersion: java.lang.String -->
   <p:dialog style="margin-left: 120px; min-width: 550px; max-width: 700px;" header="#{ivy.cm.co('/deployment/DeployToApp')} '#{appName} (v#{appVersion})'"
     id="fileUploadModal" resizable="false" widgetVar="fileUploadModal" modal="true" responsive="true"
     closeOnEscape="true">
