+++ engine-cockpit/webContent/resources/composite/ApplicationTabs.xhtml
@@ -5,7 +5,7 @@
   <cc:attribute name="tabChange"
     shortDescription="TableRowSelected will be invoked on the callback when a person is selected."
     method-signature="void handleEvent()" required="true" />
-  <cc:attribute name="update" default="" />
+  <cc:attribute name="update" default="" type="java.lang.String"/>
 </cc:interface>
 <cc:implementation>
   <p:tabView id="applicationTabView" styleClass="tab-view-card" dynamic="true"
