+++ engine-cockpit/webContent/resources/composite/Bean.xhtml
@@ -3,7 +3,7 @@
   xmlns:cc="http://xmlns.jcp.org/jsf/composite" xmlns:ic="http://ivyteam.ch/jsf/component"
   xmlns:p="http://primefaces.org/ui" xmlns:er="http://xmlns.jcp.org/jsf/composite/composite">
 <cc:interface>
-  <cc:attribute name="event" />
+  <cc:attribute name="event" type="ch.ivyteam.enginecockpit.monitor.events.Event" />
 </cc:interface>
 
 <cc:implementation>
