+++ engine-cockpit/webContent/resources/composite/BreadcrumbInfos.xhtml
@@ -4,7 +4,7 @@
   xmlns:p="http://primefaces.org/ui">
 
 <cc:interface>
-  <cc:attribute name="appContext" default="true"
+  <cc:attribute name="appContext" default="true" type="boolean"
       shortDescription="Flage indicating if the app or the security context is shown; default true = app"/>
 </cc:interface>
 <cc:implementation>
