+++ engine-cockpit/webContent/resources/composite/Configuration.xhtml
@@ -4,9 +4,9 @@
   xmlns:p="http://primefaces.org/ui" xmlns:c="http://xmlns.jcp.org/jsp/jstl/core"
   xmlns:pe="http://primefaces.org/ui/extensions">
 <cc:interface>
-  <cc:attribute name="bean" />
-  <cc:attribute name="dialogTitle" default="#{ivy.cm.co('/configuration/ConfigurationDialogDefaultTitle')}" />
-  <cc:attribute name="showContentFilter" default="true" />
+  <cc:attribute name="bean" type="ch.ivyteam.enginecockpit.configuration.model.ConfigView"/>
+  <cc:attribute name="dialogTitle" default="#{ivy.cm.co('/configuration/ConfigurationDialogDefaultTitle')}" type="java.lang.String" />
+  <cc:attribute name="showContentFilter" default="true" type="boolean" />
 </cc:interface>
 <cc:implementation>
   <h:form id="form" onkeypress="if (event.keyCode == 13) return false;">
