+++ engine-cockpit/webContent/resources/composite/Error.xhtml
@@ -4,7 +4,7 @@
   xmlns:p="http://primefaces.org/ui">
 <cc:interface>
   <cc:attribute name="error" type="ch.ivyteam.enginecockpit.util.ErrorValue"/>
-  <cc:attribute name="styleClass"/>
+  <cc:attribute name="styleClass" type="java.lang.String" />
 </cc:interface>
 <cc:implementation>
   <h:panelGroup styleClass="flex justify-content-between #{cc.attrs.styleClass}">
