+++ engine-cockpit/webContent/resources/composite/HelpServiceDialog.xhtml
@@ -3,7 +3,7 @@
   xmlns:cc="http://xmlns.jcp.org/jsf/composite" xmlns:ic="http://ivyteam.ch/jsf/component"
   xmlns:p="http://primefaces.org/ui" xmlns:pe="http://primefaces.org/ui/extensions">
 <cc:interface>
-  <cc:attribute name="service" required="true" />
+  <cc:attribute name="service" required="true" type="ch.ivyteam.enginecockpit.services.help.HelpServices" />
 </cc:interface>
 <cc:implementation>
   <p:dialog styleClass="config-showfile" header="#{cc.attrs.service.title}" id="helpServicesModal"
