+++ engine-cockpit/webContent/resources/composite/OverviewBox.xhtml
@@ -3,11 +3,11 @@
   xmlns:cc="http://xmlns.jcp.org/jsf/composite" xmlns:ic="http://ivyteam.ch/jsf/component"
   xmlns:p="http://primefaces.org/ui">
 <cc:interface>
-  <cc:attribute name="box" />
-  <cc:attribute name="icon" />
-  <cc:attribute name="title" />
-  <cc:attribute name="count" />
-  <cc:attribute name="link" />
+  <cc:attribute name="box" type="java.lang.String"/>
+  <cc:attribute name="icon" type="java.lang.String"/>
+  <cc:attribute name="title" type="java.lang.String"/>
+  <cc:attribute name="count" type="java.lang.String"/>
+  <cc:attribute name="link" type="java.lang.String"/>
 </cc:interface>
 <cc:implementation>
   <div class="col">
