+++ engine-cockpit/webContent/resources/composite/PropertyEditor.xhtml
@@ -4,10 +4,11 @@
   xmlns:p="http://primefaces.org/ui"
   xmlns:jc="http://xmlns.jcp.org/jsf/composite/demo-jsfcomponents">
 <cc:interface>
-  <cc:attribute name="bean" />
-  <cc:attribute name="formName" />
-  <cc:attribute name="isNewProperty" />
+  <cc:attribute name="bean" type="ch.ivyteam.enginecockpit.services.DetailBean"/>
+  <cc:attribute name="formName" type="java.lang.String"/>
+  <cc:attribute name="isNewProperty" type="boolean"/>
 </cc:interface>
+<!-- property: ch.ivyteam.enginecockpit.commons.Property -->
 <cc:implementation>
   <p:commandButton id="editPropertyBtn"
     actionListener="#{cc.attrs.bean.setProperty(property.name)}"
