+++ engine-cockpit/webContent/resources/composite/Sidebar.xhtml
@@ -4,7 +4,7 @@
   xmlns:p="http://primefaces.org/ui">
 <cc:interface>
   <cc:attribute name="icon" default="analytics-graph-line" />
-  <cc:attribute name="jmxSource" default="" />
+  <cc:attribute name="jmxSource" default="" type="java.lang.String" />
 </cc:interface>
 
 <cc:implementation>
