+++ engine-cockpit/webContent/resources/composite/SoapHistory.xhtml
@@ -3,7 +3,7 @@
   xmlns:cc="http://xmlns.jcp.org/jsf/composite" xmlns:ic="http://ivyteam.ch/jsf/component"
   xmlns:p="http://primefaces.org/ui">
 <cc:interface>
-  <cc:attribute name="bean" required="true" />
+  <cc:attribute name="bean" required="true" type="ch.ivyteam.enginecockpit.services.WebserviceDetailBean" />
 </cc:interface>
 <cc:implementation>
   <div class="card">
