+++ engine-cockpit/webContent/resources/composite/SslTable.xhtml
@@ -6,8 +6,8 @@
   xmlns:ic="http://ivyteam.ch/jsf/component"
   xmlns:p="http://primefaces.org/ui">
 <cc:interface>
-  <cc:attribute name="bean" required="true" />
-  <cc:attribute name="title" required="true" />
+  <cc:attribute name="bean" required="true" type="ch.ivyteam.enginecockpit.system.ssl.KeyStoreBean" />
+  <cc:attribute name="title" required="true" type="java.lang.String" />
 </cc:interface>
 
 <cc:implementation>
