+++ engine-cockpit/webContent/resources/composite/SystemDb.xhtml
@@ -3,7 +3,7 @@
   xmlns:cc="http://xmlns.jcp.org/jsf/composite" xmlns:ic="http://ivyteam.ch/jsf/component"
   xmlns:p="http://primefaces.org/ui">
 <cc:interface>
-  <cc:attribute name="updateFields" default="systemDb:systemDbForm:connectionPanel" />
+  <cc:attribute name="updateFields" default="systemDb:systemDbForm:connectionPanel" type="java.lang.String" />
 </cc:interface>
 
 <cc:implementation>
