+++ engine-cockpit/webContent/resources/composite/WebServer.xhtml
@@ -3,8 +3,8 @@
   xmlns:cc="http://xmlns.jcp.org/jsf/composite" xmlns:ic="http://ivyteam.ch/jsf/component"
   xmlns:p="http://primefaces.org/ui">
 <cc:interface>
-  <cc:attribute name="updateFields" default="webServer:webserverWarnMessage" />
-  <cc:attribute name="responsiveClasses" default="col-12 md:col-6 lg:col-6" />
+  <cc:attribute name="updateFields" default="webServer:webserverWarnMessage" type="java.lang.String"/>
+  <cc:attribute name="responsiveClasses" default="col-12 md:col-6 lg:col-6" type="java.lang.String"/>
 </cc:interface>
 
 <cc:implementation>
