+++ engine-cockpit/webContent/resources/demo-jsfcomponents/TlsTesterMissingCertView.xhtml
@@ -6,7 +6,7 @@
   xmlns:ic="http://ivyteam.ch/jsf/component"
   xmlns:p="http://primefaces.org/ui">
 <cc:interface>
-  <cc:attribute name="bean" required="true" />
+  <cc:attribute name="bean" required="true" type="ch.ivyteam.enginecockpit.system.ssl.TlsTesterBean"/>
   <cc:attribute name="cert" required="false" type="java.security.cert.X509Certificate" />
 </cc:interface>
