+++ engine-cockpit/webContent/view/monitorClassHistogram.xhtml
@@ -2,7 +2,7 @@
   xmlns:h="http://xmlns.jcp.org/jsf/html" xmlns:ui="http://xmlns.jcp.org/jsf/facelets"
   xmlns:ic="http://ivyteam.ch/jsf/component" xmlns:p="http://primefaces.org/ui"
   xmlns:pe="http://primefaces.org/ui/extensions" xmlns:cc="http://xmlns.jcp.org/jsf/composite/composite"
-  xmlns:c="http://java.sun.com/jsp/jstl/core">
+  xmlns:c="http://xmlns.jcp.org/jsp/jstl/core">
   <h:body>
     <ui:composition template="../includes/template/template.xhtml">
       <ui:define name="breadcrumb">
